Move authControllerSpy to it's own file
