+++ TirelessTests/Auth/Shared/QuickSignInViewModelTests.swift
@@ -6,7 +6,6 @@
 //
 
 import XCTest
-import Combine
 import Tireless
 
 class QuickSignInViewModelTests: XCTestCase {
@@ -72,28 +71,6 @@ class QuickSignInViewModelTests: XCTestCase {
         let httpSpy = HTTPClientSpy()
         let sut = QuickSignInViewModel(appleServices: authSpy, firestore: httpSpy)
         return (sut, authSpy, httpSpy)
-    }
-    
-    
-    private class AuthControllerSpy: AuthController {
-        
-        private(set) var authenticateCallCount = 0
-        private var authenticateCompletions = [PassthroughSubject<AuthData, AuthError>]()
-        
-        func authenticate() -> AnyPublisher<AuthData, AuthError> {
-            let publisher = PassthroughSubject<AuthData, AuthError>()
-            authenticateCompletions.append(publisher)
-            authenticateCallCount += 1
-            return publisher.eraseToAnyPublisher()
-        }
-        
-        func completeAuthenticate(with error: AuthError, at index: Int = 0) {
-            authenticateCompletions[index].send(completion: .failure(error))
-        }
-        
-        func completeAuthenticateSuccessfully(with data: AuthData, at index: Int = 0) {
-            authenticateCompletions[index].send(data)
-        }
     }
     
     private class HTTPClientSpy: HTTPClient {
